Fix warning that $start is not available in footer.php
Most of the time, the footer is able to read the `$start` variable declared in the header, but sometimes this doesn't work.
Add a global statement to explicitly say that `$start` is part of the global namespace before trying to modify/read it.
In case `$start` is not available or not set to a non-zero value, do not try to print the page generation time.